Got it working with database connection cleaned up files, refactored
