+++ src/main/java/com/runmvc/JsonApi/JsonApiDataDTO.java
@@ -1,4 +1,4 @@
-package com.runmvc.JsonApi;
+package com.runmvc.jsonApi;
 
 import java.util.Map;
