+++ src/main/java/com/runmvc/PersonRepository.java
@@ -1,0 +1,11 @@
+package com.runmvc;
+
+import com.runmvc.entities.Person;
+import org.springframework.data.repository.CrudRepository;
+
+/**
+ * Created by johnjastrow on 4/6/17.
+ */
+public interface PersonRepository extends CrudRepository<Person, Integer> {
+
+}
