+++ src/main/resources/application.properties
@@ -1,12 +1,13 @@
-server.port = 8080
+#server.port = 8080
 spring.datasource.url=jdbc:postgresql://localhost:5432/runs
-spring.datasource.username=ryan
-spring.datasource.password=
-spring.jpa.hibernate.ddl-auto=update
-
-# spring.datasource.max-active=5
-# spring.datasource.max-idle=1
-# spring.datasource.min-idle=1
-# spring.datasource.initial-size=1
-# spring.datasource.testOnBorrow=true
-spring.datasource.validationQuery=SELECT 1
+spring.jpa.generate-ddl=true
+#spring.datasource.username=
+#spring.datasource.password=
+#spring.jpa.hibernate.ddl-auto=update
+#
+## spring.datasource.max-active=5
+## spring.datasource.max-idle=1
+## spring.datasource.min-idle=1
+## spring.datasource.initial-size=1
+## spring.datasource.testOnBorrow=true
+#spring.datasource.validationQuery=SELECT 1
